create submission service (parent of modeling and text submission service) and move common method for checking the submission lock limit there
